fix(mirror): fix relative positioning
remove sign inversion
use double for more accurate calculation.
clamp values to available range.

+++ main_board/src/optics/mirror/mirror_private.c
@@ -64,7 +64,7 @@ const int32_t mirror_center_angles[MOTORS_COUNT] = {
     [MOTOR_THETA_ANGLE] = MIRROR_ANGLE_THETA_CENTER_MILLIDEGREES,
     [MOTOR_PHI_ANGLE] = MIRROR_ANGLE_PHI_CENTER_MILLIDEGREES};
 
-const uint32_t microsteps_per_mm = 12800; // 1mm / 0.4mm (pitch) * (360° / 18°
+const double microsteps_per_mm = 12800.0; // 1mm / 0.4mm (pitch) * (360° / 18°
 // (per step)) * 256 micro-steps
 
 // SPI w/ TMC5041
@@ -86,25 +86,25 @@ static struct spi_buf_set tx_bufs = {.buffers = &tx, .count = 1};
 
 int32_t
 calculate_millidegrees_from_center_position(
-    int32_t microsteps_from_center_position, const float motors_arm_length_mm)
+    int32_t microsteps_from_center_position, const double motors_arm_length_mm)
 {
-    float stepper_position_from_center_millimeters =
-        (float)microsteps_from_center_position / (float)microsteps_per_mm;
+    double stepper_position_from_center_millimeters =
+        (double)microsteps_from_center_position / microsteps_per_mm;
     int32_t angle_from_center_millidegrees =
-        asinf(stepper_position_from_center_millimeters / motors_arm_length_mm) *
-        180000.0f / M_PI;
+        asin(stepper_position_from_center_millimeters / motors_arm_length_mm) *
+        180000.0 / M_PI;
     return angle_from_center_millidegrees;
 }
 
 int32_t
 calculate_microsteps_from_center_position(
-    int32_t angle_from_center_millidegrees, const float motors_arm_length_mm)
+    int32_t angle_from_center_millidegrees, const double motors_arm_length_mm)
 {
-    float stepper_position_from_center_millimeters =
-        sinf((float)angle_from_center_millidegrees * M_PI / 180000.0f) *
+    double stepper_position_from_center_millimeters =
+        sin((double)angle_from_center_millidegrees * M_PI / 180000.0) *
         motors_arm_length_mm;
-    int32_t stepper_position_from_center_microsteps = (int32_t)roundf(
-        stepper_position_from_center_millimeters * (float)microsteps_per_mm);
+    int32_t stepper_position_from_center_microsteps = (int32_t)roundl(
+        stepper_position_from_center_millimeters * microsteps_per_mm);
     return stepper_position_from_center_microsteps;
 }
